Move executor instantiation from `cirrus-executor` into `cirrus-service`
