+++ cumulus/client/cirrus-service/Cargo.toml
@@ -29,6 +29,7 @@ tracing = "0.1.32"
 # Cirrus
 cirrus-client-executor = { path = "../cirrus-executor" }
 cirrus-client-executor-gossip = { path = "../executor-gossip" }
+cirrus-node-primitives = { path = "../../../crates/cirrus-node-primitives" }
 cirrus-primitives = { path = "../../primitives" }
 
 # Subspace
